Fills missing date values in a time series

Fills missing date values in a time series

Fills missing date values in a time series

fill_missing_dates(df, date_col, id_cols = NULL, frequency = NULL)

fill_missing_dates(df, date_col, id_cols = NULL, frequency = NULL)

Arguments

df
A data frame representing one or more time series of the same frequency.

date_col
A character indicating the date column in the data frame.

id_cols
A character vector of column names of "df"; this argument is only necessary if passing multiple time series into the function; the id_cols uniquely identify each time series.

frequency
The frequency of the time series; one of "day", "week", "month", "quarter", "year", If left NULL, will attempt to auto-calculate.

Examples

{
data(economics)
set.seed(12345)

# Fill in missing data for a single time series
test_data = economics %>% select(date, pce) %>% sample_n(10)
fill_missing_dates(test_data, 'date', frequency = 'month')

# Fill in missing data for multiple time series
test_data = economics %>%
 sample_n(100) %>%
 pivot_longer(., -date, names_to = 'varname')
fill_missing_dates(test_data, 'date', id_cols = 'varname', frequency = 'month')

}

#> Warning: data set 'economics' not found

#> Error in select(., date, pce): object 'economics' not found

{
data(economics)
set.seed(12345)

# Fill in missing data for a single time series
test_data = economics %>% select(date, pce) %>% sample_n(10)
fill_missing_dates(test_data, 'date', frequency = 'month')

# Fill in missing data for multiple time series
test_data = economics %>%
 sample_n(100) %>%
 pivot_longer(., -date, names_to = 'varname')
fill_missing_dates(test_data, 'date', id_cols = 'varname', frequency = 'month')

}

#> Warning: data set 'economics' not found

#> Error in select(., date, pce): object 'economics' not found